Выносит общую логику обновления
